SW: Fix warning: type does not match original declaration [enabled by default]
git-svn-id: https://svn.eduke32.com/eduke32@5226 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/sw/src/copysect.c
@@ -163,7 +163,7 @@ void CopySectorMatch(short match)
                         if (TEST(sector[dest_sp->sectnum].extra, SECTFX_SECTOR_OBJECT))
                         {
                             SECTOR_OBJECTp sop;
-                            extern short GlobSpeedSO;
+                            extern int GlobSpeedSO;
 
                             // find and add sprite to SO
                             sop = DetectSectorObject(&sector[sprite[src_move].sectnum]);
